fix(gsd): align reopen drift checks
