+++ src/resources/extensions/gsd/state-reconciliation/drift/artifact-db.ts
@@ -22,7 +22,6 @@ import { clearParseCache } from "../../files.js";
 import {
   clearPathCache,
   gsdProjectionRoot,
-  gsdRoot,
   resolveMilestonePath,
   resolveSliceFile,
   resolveTaskFile,
@@ -30,7 +29,7 @@ import {
 import { isClosedStatus } from "../../status-guards.js";
 import { invalidateStateCache } from "../../state.js";
 import type { GSDState } from "../../types.js";
-import { readEvents } from "../../workflow-events.js";
+import { isAfter, latestExplicitReopenAt } from "../../milestone-reopen-events.js";
 import type { DriftContext, DriftHandler, DriftRecord } from "../types.js";
 
 type DiskSliceIdDivergenceDrift = Extract<
@@ -99,30 +98,6 @@ function latestCompletedMilestoneDispatch(
   } catch {
     return null;
   }
-}
-
-function latestExplicitReopenAt(basePath: string, milestoneId: string): string | null {
-  const root = gsdRoot(basePath);
-  const candidates = [
-    join(root, "event-log.jsonl"),
-    join(root, `event-log-${milestoneId}.jsonl.archived`),
-  ];
-
-  let latest: string | null = null;
-  for (const file of candidates) {
-    for (const event of readEvents(file)) {
-      const eventMilestoneId = (event.params as { milestoneId?: unknown }).milestoneId;
-      if (event.cmd !== "reopen-milestone" || eventMilestoneId !== milestoneId) continue;
-      if (!latest || event.ts > latest) latest = event.ts;
-    }
-  }
-  return latest;
-}
-
-function isAfter(value: string | null | undefined, cutoff: string | null): boolean {
-  if (!cutoff) return true;
-  if (!value) return true;
-  return Date.parse(value) > Date.parse(cutoff);
 }
 
 function hasExplicitReopenAfter(
